requirements.txt
# =========================
# LangChain – Compatible & Existing Versions
# =========================
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-text-splitters==0.2.2

# langchain-openai version that matches core 0.2.38
langchain-openai==0.1.23

# =========================
# OpenAI SDK (Azure works through this)
# =========================
openai==1.40.6
# =========================
# Vector Database
# =========================
faiss-cpu==1.8.0

# =========================
# API Server
# =========================
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# =========================
# UI
# =========================
streamlit==1.36.0

# =========================
# Document Processing
# =========================
pypdf==4.2.0
python-docx==1.1.2
docx2txt==0.8

# =========================
# Utilities
# =========================
python-dotenv==1.0.1
tiktoken==0.7.0

# =========================
# Pydantic v1 (required by LangChain 0.2.x)
# =========================
pydantic==1.10.13
